Move audio.js to demo folder
